==> scripts/00_libraries.R <==
# load libraries ---------------------------

library("tidyverse")
library("papaja")
library("knitr")
library("here")
library("dplyr")
library("ggplot2")
library("lme4")
library("lmerTest")
library("psych")
#library(readtextgrid)
library("tidyverse")
library("papaja")
library("knitr")
library("here")
library("fs")
library("purrr")
library("dplyr")
library("tidyr")
library("forcats")
library("readr")
library("readxl")
library("writexl")
library("glue")
library("ggplot2")
library("ganttrify")
#library("magick")
library("webshot")
library("kableExtra")
library("hrbrthemes")
library(scales)
library(HDInterval) 
# ------------------------------------------
